requirements.txt
pytorch-lightning
itk
numpy
nibabel
